+++ src/main/kotlin/com/github/yusufugurozbek/testcontainers/port/updater/DatasourceUrlExtractor.kt
@@ -2,7 +2,7 @@ package com.github.yusufugurozbek.testcontainers.port.updater
 
 class DatasourceUrlExtractor {
 
-    private val regex: Regex = "Database: (.*?) \\((.*?)\\)".toRegex()
+    private val regex: Regex = "Database: (.*?)(?!\\S)".toRegex()
 
     internal fun extract(from: String): String? {
         val find = regex.find(from)