+++ src/test/kotlin/com/github/yusufugurozbek/testcontainers/port/updater/DatasourceUrlExtractorTest.kt
@@ -1,7 +1,9 @@
 package com.github.yusufugurozbek.testcontainers.port.updater
 
-import com.github.yusufugurozbek.testcontainers.port.updater.TestConstants.DATASOURCE_URL
-import com.github.yusufugurozbek.testcontainers.port.updater.TestConstants.LOG_ENTRY
+import com.github.yusufugurozbek.testcontainers.port.updater.TestConstants.DATASOURCE_URL_1
+import com.github.yusufugurozbek.testcontainers.port.updater.TestConstants.DATASOURCE_URL_2
+import com.github.yusufugurozbek.testcontainers.port.updater.TestConstants.LOG_ENTRY_1
+import com.github.yusufugurozbek.testcontainers.port.updater.TestConstants.LOG_ENTRY_2
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 
@@ -10,8 +12,13 @@ internal class DatasourceUrlExtractorTest {
     private val sut = DatasourceUrlExtractor()
 
     @Test
-    fun `extractDataSourceUrl successfully extracts url if present`() {
-        assertEquals(sut.extract(LOG_ENTRY), DATASOURCE_URL)
+    fun `extractDataSourceUrl successfully extracts url if present with driver part`() {
+        assertEquals(sut.extract(LOG_ENTRY_1), DATASOURCE_URL_1)
+    }
+
+    @Test
+    fun `extractDataSourceUrl successfully extracts url if present without driver part`() {
+        assertEquals(sut.extract(LOG_ENTRY_2), DATASOURCE_URL_2)
     }
 
     @Test